+++ CHANGELOG.md
@@ -5,5 +5,35 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Default output directory.
 
-[Unreleased]: https://github.com/ferraith/setuptools-pyecore/compare/943d95a...HEAD
+## [0.2.0] - 2018-09-03
+### Added
+- Sanity checks for user input.
+- Logging messages.
+
+### Changed
+- `Overview` section in README file.
+
+### Fixed
+- Faulty interpretation of user module path.
+
+## [0.1.1] - 2018-08-29
+### Fixed
+- Output directory of `library` sample project.
+- Install requirements of `setuptools-pyecore`.
+- Removal of a resource from a resource set if loading failed.
+
+## [0.1.0] - 2018-08-26
+### Added
+- Command line user interface.
+- Discovery of Ecore models.
+- Code generation from Ecore models.
+- `library` sample project.
+- Reference config file.
+
+[Unreleased]: https://github.com/pyecore/setuptools-pyecore/compare/0.2.0...HEAD
+[0.2.0]: https://github.com/pyecore/setuptools-pyecore/compare/0.1.1...0.2.0
+[0.1.1]: https://github.com/pyecore/setuptools-pyecore/compare/0.1.0...0.1.1
+[0.1.0]: https://github.com/pyecore/setuptools-pyecore/compare/683af85...0.1.0